Database & Client Deliverables
Ian and Adam created a DB for us based on this zip file we sent them.
Use the main(2).wire to try out the DB.  Obviously, it will need to be
renamed "main.wire" and ran in a new "app" in RW. Also, you will need to
install the Firefox plugin Adam set us.  I resent it to you in an email
dated 4/28/2013.